requirements.txt
Flask==2.3.2
line-bot-sdk==2.4.1
openai==0.27.0